pyproject.toml
[project]
name = "reckonsys-gen-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "backoff>=2.2.1",
    "bs4>=0.0.2",
    "mcp[cli]>=1.20.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "tavily-python>=0.7.12",
]
